# pcsx2: Improve GameDB handling