--- a/common/build/Utilities/utilities.vcxproj
+++ b/common/build/Utilities/utilities.vcxproj
@@ -67,7 +67,6 @@
     <ClCompile Include="..\..\src\Utilities\Console.cpp" />
     <ClCompile Include="..\..\src\Utilities\Exceptions.cpp" />
     <ClCompile Include="..\..\src\Utilities\FastFormatString.cpp" />
-    <ClCompile Include="..\..\src\Utilities\HashTools.cpp" />
     <ClCompile Include="..\..\src\Utilities\IniInterface.cpp" />
     <ClCompile Include="..\..\src\Utilities\pxStreams.cpp" />
     <ClCompile Include="..\..\src\Utilities\pxTranslate.cpp" />
@@ -121,7 +120,6 @@
     <ClInclude Include="..\..\include\Utilities\Exceptions.h" />
     <ClInclude Include="..\..\include\Utilities\FixedPointTypes.h" />
     <ClInclude Include="..\..\include\Utilities\General.h" />
-    <ClInclude Include="..\..\include\Utilities\HashMap.h" />
     <ClInclude Include="..\..\include\Utilities\MathUtils.h" />
     <ClInclude Include="..\..\include\Utilities\MakeUnique.h" />
     <ClInclude Include="..\..\include\Utilities\MemcpyFast.h" />