--- a/common/src/Utilities/CMakeLists.txt
+++ b/common/src/Utilities/CMakeLists.txt
@@ -25,7 +25,6 @@ set(UtilitiesSources
 	EventSource.cpp
 	Exceptions.cpp
 	FastFormatString.cpp
-	HashTools.cpp
 	IniInterface.cpp
 	Linux/LnxHostSys.cpp
 	Mutex.cpp
@@ -57,7 +56,6 @@ set(UtilitiesHeaders
 	../../include/Utilities/Exceptions.h
 	../../include/Utilities/FixedPointTypes.h
 	../../include/Utilities/General.h
-	../../include/Utilities/HashMap.h
 	../../include/Utilities/MakeUnique.h
 	../../include/Utilities/MemcpyFast.h
 	../../include/Utilities/MemsetFast.inl