… common/include/Utilities/HashMap.h
… common/src/Utilities/HashTools.cpp