Changes:
 * Remove GameDB write code - I removed the dialog, so nothing uses this code now.
 * Simplify Game_Data member functions - The overloads were kinda unnecessary and it fixes an infinite recursion warning on FreeBSD (basically, wxWidgets 3 wxString conversion fun).
 * Simplify hashing - There's no point in making everything lowercase when the actual key is case-sensitive, and we can just use std::hash instead. It improves load time a bit (more so on Windows than on Linux) and fixes a FreeBSD compile error.
 * Remove the Blocktable - it hardly improves performance compared to directly using the unordered_map. The blocktable was only marginally better on Windows and it made no difference on Linux (and before I removed the lowercase conversion in the hash function the blocktable was more than marginally worse on Windows). Note that this slightly changes the behaviour if there are duplicate serial entries in the GameDB (current behaviour is to use only the last duplicate entry, PR behaviour is to merge the entries, with the last duplicate entry replacing any fields that are present in previous duplicate entries). Though this shouldn't matter much because we shouldn't really be keeping/introducing any duplicate GameDB entries.
 * Don't store the GameDB serial in the Game_Data key list - it's already stored in the id member variable, so there's no real need to store it again. It improves load time a little on Windows (didn't notice Linux improvements).
 * Remove HashMap from Utilities - It's now unused due to the hashing change.

On my Windows system the GameDB load time on a VS2017 release build goes from ~134ms to ~118ms (debug build went from something like ~2200-2300ms to ~1800ms). On Linux the load time goes from ~132ms to ~128ms.