@avih 2018-08-27
Nice.

> Simplify hashing - There's no point in making everything lowercase when the actual key is case-sensitive

Not sure I understand this one. Does this mean the serial (or CRC?) is now case sensitive at `GameIndex.dbf` ?