@turtleli 2018-08-27
The serial is case sensitive, but it's like that on master as well.