@turtleli 2018-08-27
I should probably provide a better explanation.

Essentially, I have no clue why it's converting the string to lowercase in the hash function. It increases the hashing time but I don't see what for. If the intention was to make the key lookup case-insensitive then no, it doesn't do that.