@avih 2018-08-27
Please be patient with me as I don't recall the details.

A hash is usually used when you want a fast access to some already-stored data by some key.

I'm guessing the stored data is the `GameIndex.dbf` items and the keys are the serials, yes?

When is it being accessed? and where is the access key taken from while accessing it by key?