@turtleli 2018-08-27
> I'm guessing the stored data is the GameIndex.dbf items and the keys are the serials, yes?

Yes.

> When is it being accessed?  and where is the access key taken from while accessing it by key?

The data is accessed when loading a game. The key is provided by the GameDB file when creating the database or from the disc serial (ELF filename modified into serial form) when loading a game.

Actually, now that I think about it, what was probably missing was a case insensitive equality operator like
```
struct StringCompareNoCase
{
    bool operator()(const wxString &lhs, const wxString &rhs) const
    {
        return lhs.CmpNoCase(rhs) == 0;
    }
}
```
which actually would have required a case insensitive hash. I assume that's what was intended, otherwise the lowercase really makes no sense.