@turtleli 2018-08-27
> Right, so performance isn't really a concern here,

Yes, main thing was reducing the compile errors/warnings on FreeBSD, but I got sidetracked questioning why some of the stuff was necessary...

> it seems the one at GameIndex.dbf was converted to lower case.
> Is it possible that the ELF-derived serial is already converted to lower case elsewhere?
> Otherwise, if the games db serial is converted and the ELF-derived one isn't, then there would have never been any matches... (unless both key sources only had lower-case string to begin with).

No. Both serials are uppercase. The hash function in master calculates the hash based on the lower case version of the serial but doesn't modify the serial itself.

The hash is used only for the first part of the lookup to find the bucket the data is in (if it exist). The key is then compared with the other keys in the bucket to find a match based on what the equality criteria is.

Converting the serial to lowercase in the hash only changes what bucket the data is accessed/stored in. If the serial comparison was case insensitive then that would make sense as it'd put all upper/lower/mixed case variations of the serial into the same bucket and would prevent duplicate keys.