@avih 2018-08-28
Thanks for the explanation.

> The hash function in master calculates the hash based on the lower case version

"in master" means before this patchset, yes?

Anyway, yes, that's what I meant, though I didn't realize it's used both converted and in original case:

> The hash is used only for the first part of the lookup to find the bucket the data is in (if it exist). The key is then compared with the other keys in the bucket to find a match based on what the equality criteria is.
>
> Converting the serial to lowercase in the hash only changes what bucket the data is accessed/stored in

Which I find a bit weird. I agree that if lower case is only used for locating the bucket then it's indeed useless.

I also assume that all the serials both at `GameIndex.dbf` and the ones derived from the ELF are all upper case, though that might require some closer inspection, at least of the DB.

Is there any reason not to treat all serials as case-less (specifically convert all to lower case both for the task of choosing a bucket and for the final comparison)?

It seems to me that completely ignoring the case would be more resilient to user errors at the DB, and I find it extremely hard to believe that two serials which differ only by case actually refer to genuinely different disks, right?

And who knows, it might fix some lower-case serial at the DB which were never matched till now ;)