@turtleli 2018-08-28
> "in master" means before this patchset, yes?

Yes.

> Is there any reason not to treat all serials as case-less (specifically convert all to lower case both for the task of choosing a bucket and for the final comparison)?
> 
> It seems to me that completely ignoring the case would be more resilient to user errors at the DB, and I find it extremely hard to believe that two serials which differ only by case actually refer to genuinely different disks, right?

There is a good reason. I would have to redo my commits. ;)

Nah, it's probably fine to make it case insensitive unless someone else knows a reason for why it should be case sensitive. I'll update the PR soon(ish).