@avih 2018-08-28
> There is a good reason. I would have to redo my commits

Actually, I also find it a good reason. It also keeps the code simpler, and doesn't take away any robustness (which the original code didn't have anyway).

I'd be fine with just taking the serials as is everywhere.

I'm guessing that if someone adds a serial to the DB, it's their responsibility to make sure that it's actually recognized when running the game disk. It worked so far ;)