@turtleli 2018-08-28
Well, it's not really much more complex, but yeah, I guess having people make sure any serials they add are in uppercase shouldn't be asking too much of them.

Update aborted. ;)