@turtleli 2018-08-29
Updated to increase minimum bucket count. No other changes.

Will merge after buildbots are done.